Separated create/update/delete functionality from ExhibitPagesController, so that functionality can be reused by other controllers, fixed ExhibitsController (pages should now be correctly deleted when deleting an exhibit)
