+++ HiP-DataStore/Controllers/ExhibitsController.cs
@@ -4,6 +4,7 @@ using MongoDB.Driver;
 using PaderbornUniversity.SILab.Hip.DataStore.Core;
 using PaderbornUniversity.SILab.Hip.DataStore.Core.ReadModel;
 using PaderbornUniversity.SILab.Hip.DataStore.Core.WriteModel;
+using PaderbornUniversity.SILab.Hip.DataStore.Core.WriteModel.Commands;
 using PaderbornUniversity.SILab.Hip.DataStore.Model;
 using PaderbornUniversity.SILab.Hip.DataStore.Model.Entity;
 using PaderbornUniversity.SILab.Hip.DataStore.Model.Events;
@@ -166,11 +167,7 @@ namespace PaderbornUniversity.SILab.Hip.DataStore.Controllers
             if (_referencesIndex.IsUsed(ResourceType.Exhibit, id))
                 return BadRequest(ErrorMessages.ResourceInUse);
 
-            var ev = new ExhibitDeleted { Id = id };
-            await _eventStore.AppendEventAsync(ev);
-            await RemoveExhibitReferencesAsync(id);
-
-            // TODO: Delete all pages belonging to the exhibit (cascading deletion)
+            // pages should be deleted along with the exhibit (cascading deletion) => first, remove the pages
             var pageIds = _referencesIndex.ReferencesOf(ResourceType.Exhibit, id)
                 .Where(reference => reference.Type.Name == ResourceType.ExhibitPage.Name)
                 .Select(reference => reference.Id)
@@ -178,15 +175,17 @@ namespace PaderbornUniversity.SILab.Hip.DataStore.Controllers
 
             foreach (var pageId in pageIds)
             {
-                if (!_entityIndex.Exists(ResourceType.Exhibit, id))
-                    return NotFound();
+                if (_referencesIndex.IsUsed(ResourceType.ExhibitPage, pageId))
+                    return BadRequest("The exhibit cannot be deleted because it contains pages that are referenced by other resources");
 
-                if (_referencesIndex.IsUsed(ResourceType.Exhibit, id))
-                    return BadRequest(ErrorMessages.ResourceInUse);
-
-                var pageDeleted = new ExhibitPageDeleted { Id = pageId };
-                
+                var pageDeleteEvents = ExhibitPageCommands.Delete(pageId, _referencesIndex);
+                await _eventStore.AppendEventsAsync(pageDeleteEvents);
             }
+
+            // now remove the actual exhibit
+            var ev = new ExhibitDeleted { Id = id };
+            await _eventStore.AppendEventAsync(ev);
+            await RemoveExhibitReferencesAsync(id);
 
             return NoContent();
         }
