+++ HiP-DataStore/Core/WriteModel/Commands/AddValidationErrorDelegate.cs
@@ -1,0 +1,4 @@
+﻿namespace PaderbornUniversity.SILab.Hip.DataStore.Core.WriteModel.Commands
+{
+    public delegate void AddValidationErrorDelegate(string key, string errorMessage);
+}
